mcp-sec/pyproject.toml
[tool.poetry]
name = "driftcop"
version = "0.1.0"
description = "Security scanner for MCP (Model Context Protocol) servers"
authors = ["Drift Cop Team"]
readme = "README.md"
packages = [{include = "mcp_sec", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
httpx = "^0.25.0"
pydantic = "^2.5.0"
jsonschema = "^4.20.0"
docker = "^7.0.0"
aiofiles = "^23.2.1"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"
openai = "^1.6.0"
python-Levenshtein = "^0.23.0"
sarif-om = "^1.0.4"
pyyaml = "^6.0.1"
rich = "^13.7.0"
toml = "^0.10.2"
cryptography = "^41.0.7"
sigstore = "^2.1.0"
securesystemslib = "^0.28.0"
unicodedata2 = "^15.1.0"
tree-sitter = "^0.20.4"
tree-sitter-python = "^0.20.4"
tree-sitter-javascript = "^0.20.3"
tree-sitter-typescript = "^0.20.3"
tree-sitter-java = "^0.20.2"
tree-sitter-go = "^0.20.0"
tree-sitter-rust = "^0.20.4"
tree-sitter-c-sharp = "^0.20.0"
tree-sitter-ruby = "^0.20.0"
tree-sitter-php = "^0.20.0"
tree-sitter-cpp = "^0.20.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"

[tool.poetry.scripts]
driftcop = "mcp_sec.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true